Replace aliveness/tools library with /crates one
