+++ crates/aliveness/Cargo.toml
@@ -1,14 +1,15 @@
 [package]
-name = "aliveness-client"
+name = "aliveness"
 license = "GPL-3.0-only"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-aliveness = { path = "../../tools/aliveness" }
 color-eyre = { workspace = true }
 futures-util = { workspace = true }
 hula-types = { path = "../../tools/hula/types/" }
+regex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true } 
+zbus = { workspace = true }
