+++ crates/aliveness/src/service_manager.rs
@@ -8,7 +8,7 @@ use zbus::{
     Connection, Error, Proxy,
 };
 
-#[derive(Debug, Serialize)]
+#[derive(Debug)]
 enum Service {
     Hal,
     Hula,
