+++ tools/pepsi/Cargo.toml
@@ -5,7 +5,7 @@ name = "pepsi"
 version = "2.4.1"
 
 [dependencies]
-aliveness-client = { workspace = true }
+aliveness = { workspace = true }
 bat = { workspace = true }
 clap = { workspace = true }
 clap_complete = { workspace = true }
